renaming app_trapi_1_4 to remove 1.4 references
